adding auto-generated logging configuration files
I changed the logging setup so that the notebook output isn't swamped with
logging statements. When it is displayed on Github, the focus will be on the
code and diagnostic figures, not on the thousands of lines of log statements.
These are now placed in a log file in the notebook's output folder(s). The log
file and log output level are defined in the log config file. This config file
is generated by the code in the notebook and placed in the output folder.